Refactory dataset and dataset loader
