+++ refdata/dataset/__init__.py
@@ -3,7 +3,3 @@
 # Copyright (C) 2021 New York University.
 #
 # refdata is free software; you can redistribute it and/or modify it under the
-# terms of the MIT License; see LICENSE file for more details.
-
-from refdata.loader.csv_loader import CSVLoader
-from refdata.loader.json_loader import JsonLoader
